+++ CHANGES.rst
@@ -223,6 +223,8 @@ mast
 - Deprecated ``enable_cloud_dataset`` and ``disable_cloud_dataset`` in classes where they
   are non-operational. They will be removed in a future release. [#3113]
 
+- Present users with an error when nonexistent query criteria are used in ``mast.MastMissions`` query functions. [#3126]
+
 mpc
 ^^^